+++ app/src/main/java/com/example/twoviewscompose/viewmodel/LoadingViewModel.kt
@@ -7,13 +7,13 @@ import com.example.twoviewscompose.data.models.User
 import com.example.twoviewscompose.data.repository.UserRepository
 
 class LoadingViewModel : ViewModel() {
-    private val _sum = mutableStateOf(0)
-    val sum: State<Int> get() = _sum
+    private val _sum = mutableStateOf(0L)
+    val sum: State<Long> get() = _sum
 
     private val _userList = mutableStateOf(listOf<User>())
     val userList: State<List<User>> get() = _userList
 
-    fun updateSum(number1: Int, number2: Int) {
+    fun updateSum(number1: Long, number2: Long) {
         _sum.value = number1 + number2
     }
